Rename Format3Output to Output and move directory
